Create SplitMethods folder and move SplitMethodButton to it.
