+++ src/features/expenses/AddExpense/SplitMethods/SplitMethod.styled.ts
@@ -1,0 +1,10 @@
+import { Row } from "components/Containers";
+import styled from "styled-components";
+
+export const SplitMethodButtonWrapper = styled(Row)`
+  justify-content: center;
+  align-items: center;
+  height: 64px;
+  width: 64px;
+  z-index: 100;
+`;
